Rename some files and rationalize the python library.
